+++ CHANGELOG.md
@@ -2,14 +2,16 @@
 
 ## 0.1.1
 
-* Improved output formatting and fixing some coding conventions
+* Making `pod plugins` an abstract command, with `list` the default subcommand (#11 & #12)
   [Olivier Halligon](https://github.com/AliSoftware)
-* Added `search` subcommand to search plugins by name, author and description
+* Added `search` subcommand to search plugins by name, author and description.
+  [Olivier Halligon](https://github.com/AliSoftware)
+* Refactoring, improved output formatting and fixing some coding conventions (#10)
   [Olivier Halligon](https://github.com/AliSoftware)
 
 ## 0.1.0
 
 * Initial implementation.  
   [David Grandinetti](https://github.com/dbgrandi)
-* Added `create` subcommand to create an empty project for a new plugin
+* Added `create` subcommand to create an empty project for a new plugin.
   [Boris Bügling](https://github.com/neonichu)